+++ backend/package.json
@@ -4,8 +4,8 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "dev": "nodemon src/app.js",
-    "start": "node src/app.js",
+    "dev": "nodemon app.js",
+    "start": "node app.js",
     "test": "vitest"
   },
   "author": "",
@@ -16,7 +16,7 @@
     "express": "^4.21.2",
     "firebase-admin": "^13.2.0",
     "lucide-react": "^0.510.0",
-    "mongoose": "^7.8.6",
+    "mongoose": "^7.8.7",
     "puppeteer": "^24.9.0",
     "puppeteer-extra": "^3.3.6",
     "puppeteer-extra-plugin-stealth": "^2.11.2",
