--- backend/src/app.js
@@ -1,46 +1,0 @@
-import express from "express";
-import { connectDB } from "./config/db.js";
-import userRouter from "./routes/userRoutes.js";
-import ingredientRouter from "./routes/ingredientRoutes.js";
-import recipeRouter from "./routes/recipeRoutes.js";
-import verifyToken from "./middleware/firebase-auth.js";
-import mongoose from "mongoose";
-import scrapeRouter from "./routes/scrapeRouter.js";
-const app = express();
-const PORT = process.env.PORT || 5000;
-
-connectDB()
-  .then(() => console.log("Database connected successfully"))
-  .catch((err) => console.error("Database connection error:", err));
-
-mongoose.connection.on("error", (err) => {
-  console.error("MongoDB connection error:", err);
-});
-
-mongoose.connection.once("open", () => {
-  console.log("MongoDB connected successfully");
-  mongoose.connection.db.listCollections().toArray((err, collections) => {
-    if (err) {
-      console.error("Error listing collections:", err);
-    } else {
-      console.log(
-        "Available collections:",
-        collections.map((c) => c.name)
-      );
-    }
-  });
-});
-
-app.use(express.json());
-app.get("/api/test-direct", (req, res) => {
-  res.send("Direct route working!");
-});
-
-app.use("/api/users", verifyToken, userRouter);
-app.use("/api/scrape", scrapeRouter);
-app.use("/api/ingredients", verifyToken, ingredientRouter);
-app.use("/api/recipes", verifyToken, recipeRouter);
-
-app.listen(PORT, () => {
-  console.log(`Server is running and listening on port ${PORT}`);
-});
